Fixed cursor_identities test being restored to wrong place
